requirements.txt
pyyaml>=6.0.1
requests>=2.32.0
feedparser>=6.0.11
anthropic>=0.36.0
